+++ include/swift/Runtime/Once.h
@@ -31,7 +31,7 @@ typedef long swift_once_t;
 
 // On Cygwin, std::once_flag can not be used because it is larger than the
 // platform word.
-typedef unsigned long swift_once_t;
+typedef uintptr_t swift_once_t;
 #else
 
 // On other platforms swift_once_t is std::once_flag
